backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ken-memorial-backend"
version = "0.1.0"
description = "FastAPI backend for Ken Memorial"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "PyJWT>=2.9.0",
  "psycopg[binary]>=3.2.1",
  "sqlalchemy>=2.0.35",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
